Work around pselect()/ppoll() failing to deliver handler.
In some rare cases, specially when executing long sequence of short command,
ninja does not respond to signal because its handlers are not triggered. This
happens on my linux box version 3.13-1-amd64.

This patch introduce --unprotected-syscall configure option to enable
implementation of ppoll() and pselect() based on poll() and select().
This implementation is similar to what ppoll()/pselect() does but not in an
atomic way. So theoretically there is a race condition. However since we do
not test for the "interrupted_" flag between the calls to sigprocmask() and
select()/poll() there is no race condition in practice. Most of the time
the signal is delivered when leaving sigprocmask() or when poll()/select()
returns EINTR.

See discussions:
- https://groups.google.com/forum/#!topic/ninja-build/leEnr095PeA
- http://lkml.iu.edu//hypermail/linux/kernel/1109.1/00609.html