Detach child process from tty without using setsid.
We cannot use setsid(2) because session leader process do not
receive SIGTSTP and this is needed to properly handle ctrl-z
(see #744).

See #909, #910 and #1097.

TODO: This patch was originally part of #744 but was removed when
rebasing on master because I do not know how to tell posix_spawn(2)
to perform an open(2) followed by a ioctl(2).