Suspend sub-processes when control-z is hit.
When control-z is hit the shell sends SIGTSTP to the front process group.
Typically, it suspends all the process tree started by the command. But ninja
moves its sub-processes to their own group so that it can handle
subsets of running commands. In such a case sub-processes do not
receive SIGTSTP. Only ninja is suspended and the command it has started
keep running.
This patch addresses this issue by catching SIGTSTP and forwarding to its
sub-process groups. Thus, the entire ninja's process tree is suspended as
expected.

Thanks to Peter Collingbourne <peter@pcc.me.uk>
      and Julien Tinnes <jln@chromium.org> for their help.